requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
shapely
pyproj
numpy
ezdxf
pydantic
jinja2
